ref #1173 -- revised naming to use PostQuery syntax
