+++ lib/QueryIterator.php
@@ -3,7 +3,7 @@
 namespace Timber;
 
 use Timber\Helper;
-use Timber\PostsArray;
+use Timber\PostCollection;
 
 // Exit if accessed directly
 if ( !defined('ABSPATH') ) {
@@ -69,7 +69,7 @@ class QueryIterator implements \Iterator, \Countable {
 
 	public function get_posts( $return_collection = false ) {
 		if ( isset($this->_query->posts) ) {
-			$posts = new PostsArray($this->_query->posts, $this->_posts_class);
+			$posts = new PostCollection($this->_query->posts, $this->_posts_class);
 			return ($return_collection) ? $posts : $posts->get_posts();
 		}
 	}
