+++ tests/test-timber-pagination.php
@@ -174,7 +174,7 @@ class TestTimberPagination extends Timber_UnitTestCase {
 
 	function testPostsCollectionPagination() {
 		$pids = $this->factory->post->create_many( 13 );
-		$posts = new Timber\PostCollection(array('post_type' => 'post'));
+		$posts = new Timber\PostQuery(array('post_type' => 'post'));
 		$pagination = $posts->pagination();
 		$this->assertEquals( 2, count( $pagination->pages ) );
 	}
@@ -185,7 +185,7 @@ class TestTimberPagination extends Timber_UnitTestCase {
 		$wp_rewrite->permalink_structure = false;
 		$posts = $this->factory->post->create_many( 55 );
 		$this->go_to( home_url( '?s=post' ) );
-		$posts = new Timber\PostCollection();
+		$posts = new Timber\PostQuery();
 		$pagination = $posts->pagination();
 		$this->assertEquals( user_trailingslashit(home_url().'/?s=post&#038;paged=5'), $pagination->pages[4]['link'] );
 	}
@@ -197,7 +197,7 @@ class TestTimberPagination extends Timber_UnitTestCase {
 		register_post_type( 'portfolio' );
 		$pids = $this->factory->post->create_many( 55, array( 'post_type' => 'portfolio' ) );
 		$this->go_to( home_url( '/portfolio/page/3' ) );
-		$posts = new Timber\PostCollection('post_type=portfolio&paged=3');
+		$posts = new Timber\PostQuery('post_type=portfolio&paged=3');
 		$pagination = $posts->pagination();
 		$this->assertEquals(6, count($pagination->pages));
 	}
@@ -206,7 +206,7 @@ class TestTimberPagination extends Timber_UnitTestCase {
 		$this->setPermalinkStructure('/%postname%/');
 		register_post_type( 'portfolio' );
 		$pids = $this->factory->post->create_many( 99, array( 'post_type' => 'portfolio' ) );
-		$posts = new Timber\PostCollection('post_type=portfolio&posts_per_page=20');
+		$posts = new Timber\PostQuery('post_type=portfolio&posts_per_page=20');
 		$pagination = $posts->pagination();
 		$this->assertEquals(5, count($pagination->pages));
 	}
@@ -214,9 +214,9 @@ class TestTimberPagination extends Timber_UnitTestCase {
 	function testCollectionPaginationSearchPrettyWithPostname() {
 		$this->setPermalinkStructure('/%postname%/');
 		$posts = $this->factory->post->create_many( 55 );
-		$archive = home_url( '?s=post' );
+		$archive = home_url('?s=post');
 		$this->go_to( $archive );
-		$posts = new Timber\PostCollection( 's=post' );
+		$posts = new Timber\PostQuery('s=post');
 		$pagination = $posts->pagination();
 		$this->assertEquals( 'http://example.org/page/5/?s=post', $pagination->pages[4]['link'] );
 	}
@@ -226,7 +226,7 @@ class TestTimberPagination extends Timber_UnitTestCase {
 		$posts = $this->factory->post->create_many( 55 );
 		$archive = home_url( '?s=post' );
 		$this->go_to( $archive );
-		$posts = new Timber\PostCollection( 's=post' );
+		$posts = new Timber\PostQuery('s=post');
 		$pagination = $posts->pagination();
 		$this->assertEquals( 'http://example.org/page/2/?s=post', $pagination->next['link'] );
 	}
@@ -236,7 +236,7 @@ class TestTimberPagination extends Timber_UnitTestCase {
 		$posts = $this->factory->post->create_many( 55 );
 		$archive = home_url( 'page/4/?s=post' );
 		$this->go_to( $archive );
-		$posts = new Timber\PostCollection( 's=post&paged=4' );
+		$posts = new Timber\PostQuery('s=post&paged=4');
 		$pagination = $posts->pagination();
 		$this->assertEquals( 'http://example.org/page/3/?s=post', $pagination->prev['link'] );
 	}
@@ -249,7 +249,7 @@ class TestTimberPagination extends Timber_UnitTestCase {
 		$posts = $this->factory->post->create_many( 55 );
 		$archive = home_url( '?s=post' );
 		$this->go_to( $archive );
-		$posts = new Timber\PostCollection();
+		$posts = new Timber\PostQuery();
 		$pagination = $posts->pagination();
 		$this->assertEquals( 'http://example.org/page/5/?s=post', $pagination->pages[4]['link'] );
 	}
@@ -261,7 +261,7 @@ class TestTimberPagination extends Timber_UnitTestCase {
 
 		$posts = $this->factory->post->create_many( 55 );
 		$this->go_to( home_url( '/' ) );
-		$posts = new Timber\PostCollection();
+		$posts = new Timber\PostQuery();
 		$pagination = $posts->pagination( array( 'base' => '/apricot/%_%', 'format' => 'page=%#%' ) );
 		$this->assertEquals( '/apricot/page=2', $pagination->next['link'] );
 	}
@@ -274,7 +274,7 @@ class TestTimberPagination extends Timber_UnitTestCase {
 		for($i=0; $i<30; $i++) {
 			$this->factory->post->create(array('post_title' => 'post'.$i, 'post_date' => '2014-02-'.$i));
 		}
-		$posts = new Timber\PostCollection('paged=3');
+		$posts = new Timber\PostQuery('paged=3');
 		$pagination = $posts->pagination( array( 'base' => '/apricot/%_%', 'format' => 'page=%#%' ) );
 		$this->assertEquals( '/apricot/page=2', $pagination->prev['link'] );
 	}
@@ -285,7 +285,7 @@ class TestTimberPagination extends Timber_UnitTestCase {
 		update_option( 'permalink_structure', $struc );
 		$posts = $this->factory->post->create_many( 150 );
 		$this->go_to( home_url( '/page/13' ) );
-		$posts = new Timber\PostCollection();
+		$posts = new Timber\PostQuery();
 		$expected_next_link = user_trailingslashit('http://example.org/page/14/');
 		$pagination = $posts->pagination();
 		$this->assertEquals( $expected_next_link, $pagination->next['link'] );
@@ -294,11 +294,11 @@ class TestTimberPagination extends Timber_UnitTestCase {
 	function testPostCollectionPaginationForMultiplePostTypes() {
 		register_post_type( 'recipe' );
 		$pids = $this->factory->post->create_many( 43, array( 'post_type' => 'recipe' ) );
-		$recipes = new Timber\PostCollection(array('post_type' => 'recipe'));
+		$recipes = new Timber\PostQuery(array('post_type' => 'recipe'));
 		$pagination = $recipes->pagination();
 		$this->assertEquals( 5, count( $pagination->pages ) );
 		$pids = $this->factory->post->create_many( 13 );
-		$posts = new Timber\PostCollection(array('post_type' => 'post'));
+		$posts = new Timber\PostQuery(array('post_type' => 'post'));
 		$pagination = $posts->pagination();
 		$this->assertEquals( 2, count( $pagination->pages ) );
 	}
